Chore: Introduce local-only designkit route
Another prerequisite for visual regression testing. This adds a route
available only when the dev server is running at
`/src/designkit-standalone` that loads the design kit page without
contacting the server. In order to ensure that the components in the
design kit are not reliant on store providers, we remove the card
examples that relied on the state providers to work. In addition, the
chart section is updated to use static data on load with the option to
randomize the data and fake stream it.